fix(daemon): keep rpc client import closure light
